requirements.txt
Django==5.1.6
djangorestframework==3.15.2
rest_framework_simplejwt==0.0.2
celery==5.6.2